requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0

# LLM APIs
google-generativeai>=0.3.0
cohere>=4.37
groq>=0.4.0
huggingface-hub>=0.19.0

# Text analysis
textstat>=0.7.3

# Utilities
python-dotenv>=1.0.0
